requirements.txt
Flask
python-dotenv
google-generativeai
Markdown
html5lib